Refactor: 각 라우터별 페이지 컴포넌트 기본 구조 추가
