+++ frontend/src/App.tsx
@@ -1,9 +1,12 @@
 import React, { useState } from "react";
 import { Routes, Route, useNavigate } from "react-router-dom";
 import { Header } from "./components/Header/Header";
-import Home from "./pages/home/Home";
+import Home from "./pages/home/index";
 import CalendarPage from "./pages/calendar";
 import Footer from "./components/Footer/Footer";
+import ContestPage from "./pages/contest";
+import CodingTestPage from "./pages/codingtest";
+import CertificatesPage from "./pages/certificates";
 
 const App: React.FC = () => {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
@@ -40,6 +43,9 @@ const App: React.FC = () => {
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/calendar" element={<CalendarPage />} />
+          <Route path="/contests" element={<ContestPage />} />
+          <Route path="/coding-tests" element={<CodingTestPage />} />
+          <Route path="/certificates" element={<CertificatesPage />} />
         </Routes>
       </main>
