+++ frontend/src/pages/certificates/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const CertificatesPage: React.FC = () => {
+  return (
+    <div>
+      <h1>자격증 페이지</h1>
+      {/* 자격증 리스트, 상세 등 내용이 들어갈 예정 */}
+    </div>
+  );
+};
+
+export default CertificatesPage;
