+++ frontend/src/pages/codingtest/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const CodingTestPage: React.FC = () => {
+  return (
+    <div>
+      <h1>코딩테스트 페이지</h1>
+      {/* 코딩테스트 리스트, 상세 등 내용이 들어갈 예정 */}
+    </div>
+  );
+};
+
+export default CodingTestPage;
