+++ frontend/src/pages/contest/index.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const ContestPage: React.FC = () => {
+  return (
+    <div>
+      <h1>공모전 페이지</h1>
+      {/* 공모전 리스트, 상세 등 내용이 들어갈 예정 */}
+    </div>
+  );
+};
+
+export default ContestPage;
